requirements.txt
fastapi
uvicorn
python-dotenv
langchain-groq
langchain-pinecone
pinecone-client
pinecone-plugin-inference
requests
aiohttp
